feat: Add Workspace apis

+++ src/repositories/crud.repository.ts
@@ -1,8 +1,15 @@
-import mongoose from "mongoose";
+import mongoose, { Types } from "mongoose";
+import { BadRequestError } from "../utils/api-error";
 
 export default function crudRepository<T, TAttrs = T>(
   model: mongoose.Model<T>
 ) {
+  const isValidId = (id: string | Types.ObjectId) => {
+    if (typeof id === "string") {
+      return Types.ObjectId.isValid(id);
+    }
+    return id instanceof Types.ObjectId;
+  };
   return {
     async create(data: TAttrs) {
       const doc = await model.create(data);
@@ -13,16 +20,29 @@ export default function crudRepository<T, TAttrs = T>(
       return docs;
     },
     async getById(id: string) {
+      if (!isValidId(id)) {
+        throw new BadRequestError("Invalid id");
+      }
       const doc = await model.findById(id);
       return doc;
     },
     async update(id: string, data: Partial<T>) {
+      if (!isValidId(id)) {
+        throw new BadRequestError("Invalid id");
+      }
       const doc = await model.findByIdAndUpdate(id, data, { new: true });
       return doc;
     },
     async delete(id: string) {
+      if (!isValidId(id)) {
+        throw new BadRequestError("Invalid id");
+      }
       const doc = await model.findByIdAndDelete(id);
       return doc;
+    },
+    async deleteMany(ids: string[] | Types.ObjectId[]) {
+      const docs = await model.deleteMany({ _id: { $in: ids } });
+      return docs;
     },
   };
 }
